Start organizing state machine code better
This includes pulling out a couple of utilities, resolveAllLogs and
asyncReduce, and introducing some initial JSDocs and related types, as
well as enabling TypeScript checking for supporting editors in the main
state testing files.
